% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riceigeneann.R
\name{riceigeneann}
\alias{riceigeneann}
\title{Annotation of genes of rice (Oryza Sativa)}
\description{
\code{ricegeneann} Annotation of genes of rice (Oryza Sativa). 
Input a or some gene name then return a or some information, 
including the from position, end position, position type and 
the chromosome number.
}
\author{
Xiang LI <ynaulx@gmail.com> Linna MA <973649882@qq.com>
}
